Show CMakeLists.txt for executables
When a source package declares an executable, and when this source
package has a CMakeLists.txt file, add this file as a source file to
the executable in the source build. This makes the source directory
accessible from Qt Creator's project tree and at the same time it
indicates the availability of the CMakeLists.txt.

By first opening the CMakeLists.txt from the project tree (i.e. as
a file), then explicitly opening it again (i.e. as a project)
and importing the existing build directory, the user can easily
configure a development environment for this project in Qt Creator.