Update copyright files
This is a review of exceptions from the default rules in the
copyright.in template file. New patterns were added manually for
qtbase, qtlocation, qttranslation, thus reducing the number of
per-file overrides. In qtsensors, per-file overrides were dropped
automatically due to the previous automatical drop of the unused
default mapping of src/*/doc/snippets/* to BSD-3-clause.