Handle staging and DESTDIR
This commit makes the CMAKE_STAGING_PREFIX variable available for
building packages, constructed from the INSTALL_DIR and the
CMAKE_INSTALL_PREFIX. This simplifies most packages.
Together with this change, it is now possible to use the DESTDIR
mechanism. ${DESTDIR} can be used as a placeholder for the $(DESTDIR)
Makefile variable.

A number of directory path constructions are normalized using the
sb_*_dir/prefix functions.

The external project CMakeLists.txt configuration/generation step
no longer substitutes regular CMake variable references (${NAME}).
Variables which need to be expanded before applying generator
expressions to the BUILD rules must be written as @NAME@.