android: Separate platform and compile SDK
While ANDROID_PLATFORM defines the minimum API level binaries are
expected to run on, the compile SDK defines which are the newest
features which are available to apps. Depending on the declared
target SDK of an app, it may be forced to use new APIs when running
on newer devices.

The target SDK is now set to 'android-28'. Notifications require
features introduced in API level 26 when targeting newer devices.